requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-forms-bootstrap==3.1.0
django-toolbelt==0.0.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
